backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
pydantic==2.5.2
python-dotenv==1.0.0
pymongo==4.6.0
pytesseract==0.3.10
python-docx==1.0.0
PyPDF2==3.0.1
requests==2.31.0
beautifulsoup4==4.12.2
openai==0.28.1
google-api-python-client==2.108.0
